+++ cmd/labells.go
@@ -17,28 +17,41 @@ package cmd
 import (
 	"fmt"
 
+	log "github.com/Sirupsen/logrus"
+	"github.com/keeneyetech/ecli/api"
 	"github.com/spf13/cobra"
 )
 
-// slideCmd represents the slide command
-var slideCmd = &cobra.Command{
-	Use:   "slide",
-	Short: "Manage slides",
-	Long:  `Use the slide command to upload slides, view slide info and so on.`,
+// labellsCmd represents the labells command
+var labellsCmd = &cobra.Command{
+	Use:   "ls",
+	Short: "List all existing labels",
 	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Println(cmd.UsageString())
+		labels, err := api.Labels()
+		if err != nil {
+			log.Fatal(err)
+		}
+		for _, l := range labels {
+			p := l.(map[string]interface{})
+			if !cfgQuiet {
+				fmt.Printf("%-30s %s %-25q\n", p["name"], p["color"], p["description"])
+			} else {
+				fmt.Printf("%s\n", p["name"])
+			}
+		}
 	},
 }
 
 func init() {
-	RootCmd.AddCommand(slideCmd)
+	labelCmd.AddCommand(labellsCmd)
+
 	// Here you will define your flags and configuration settings.
 
 	// Cobra supports Persistent Flags which will work for this command
 	// and all subcommands, e.g.:
-	// slideCmd.PersistentFlags().String("foo", "", "A help for foo")
+	// labellsCmd.PersistentFlags().String("foo", "", "A help for foo")
 
 	// Cobra supports local flags which will only run when this command
 	// is called directly, e.g.:
-	// slideCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
+	// labellsCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
 }